Рефакторинг пакета src/api/endpoints/admin_routers
