+++ src/api/endpoints/admin/register.py
